requirements.txt
aioredis==2.0.1
arq==0.25.0
asgiref==3.6.0
async-timeout==4.0.2
click==8.1.3
Django==4.1.7
hiredis==2.2.2
psycopg2-binary==2.9.5
redis==4.5.4
sqlparse==0.4.3
typing_extensions==4.5.0
djangorestframework==3.14.0
django-rest-knox==4.2.0